backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.8.1=pyhd8ed1ab_0
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.3.11=hecd8cb5_0
dj-database-url=2.1.0=pypi_0
django=5.0.6=pyhd8ed1ab_0
django-cors-headers=4.3.1=pyhd8ed1ab_0
djangorestframework=3.15.1=pyhd8ed1ab_0
expat=2.6.2=hcec6c5f_0
krb5=1.21.2=hb884880_0
libcxx=17.0.6=h88467a6_0
libedit=3.1.20230828=h6c40b1e_0
libexpat=2.6.2=h73e2aa4_0
libffi=3.4.4=hecd8cb5_1
libpq=16.3=h4501773_0
libsqlite=3.45.3=h92b6c6a_0
libzlib=1.2.13=h8a1eda9_5
ncurses=6.4=hcec6c5f_0
openssl=3.3.0=hd75f5a5_0
pip=24.0=py312hecd8cb5_0
psycopg2=2.9.9=py312hca9e88b_0
python=3.12.2=h9f0c242_0_cpython
python_abi=3.12=4_cp312
readline=8.2=hca72f7f_0
setuptools=69.5.1=py312hecd8cb5_0
sqlite=3.45.3=h6c40b1e_0
sqlparse=0.4.4=py312hecd8cb5_0
tk=8.6.14=h4d00af3_0
typing_extensions=4.11.0=py312hecd8cb5_0
tzdata=2024a=h04d1e81_0
wheel=0.43.0=py312hecd8cb5_0
xz=5.4.6=h6c40b1e_1
zlib=1.2.13=h8a1eda9_5
